Initial commit
The library has basic functionality, but it only promisifies
the functions within the scope that `nano` is originally called.

This means that if you call `nano` in server scope, the document
functions are not promisified properly, so you have to `require`
one version for `dbScope` and one version for `docScope`.